+++ src/main/java/com/nhnacademy/frontserver1/presentation/controller/AdminBookController.java
@@ -4,6 +4,7 @@ import com.nhnacademy.frontserver1.application.service.BookService;
 import com.nhnacademy.frontserver1.application.service.CategoryService;
 import com.nhnacademy.frontserver1.application.service.TagService;
 import com.nhnacademy.frontserver1.application.service.UploadImageService;
+import com.nhnacademy.frontserver1.common.exception.BookException;
 import com.nhnacademy.frontserver1.presentation.dto.request.book.CreateBookRequest;
 import com.nhnacademy.frontserver1.presentation.dto.request.book.UpdateBookRequest;
 import com.nhnacademy.frontserver1.presentation.dto.response.book.*;
@@ -115,5 +116,14 @@ public class AdminBookController {
         model.addAttribute("bookList", bookList);
 
         return "admin/product/admin-book-search";
+    }
+
+    @ExceptionHandler(BookException.class)
+    public String bookException(Model model) {
+
+        model.addAttribute("message", "이미 존재하는 책입니다.");
+        model.addAttribute("url", "/admin/product");
+
+        return "message";
     }
 }
